+++ src/CTBX.ImportHolidays.Backend/ImportHolidaysFeatureRegistration.cs
@@ -10,26 +10,15 @@ public static class ImportHolidaysFeatureRegistration
 {
     public static void RegisterServices(IServiceCollection services, IConfiguration configuration)
     {
-        //var connectionString = configuration
-        //                    .GetConnectionString("ctbx-events-db")
-        //                    !.GuardAgainstNullOrEmpty("ctbx-events-db");
-
         services.AddScoped<HolidaysImporter>();
         services.AddHostedService<ImportHolidaysDbSeeder>();
-        services.AddScoped<ReadCommandHandler>();
         services.Configure<FileUploadOptions>(configuration.GetSection(nameof(FileUploadOptions)));
-        //services.Configure<HolidayImporterOptions>(configuration.GetSection("POSTGRES_DB"));
-        //services.Configure<HolidayImporterOptions>(configuration
-        //                    .GetConnectionString("ctbx-events-db"));
+
         services.Configure<HolidayImporterOptions>(options =>
         { options.ConnectionString = configuration
                             .GetConnectionString("ctbx-common-db")
                             !.GuardAgainstNullOrEmpty("ctbx-common-db"); });
         services.AddScoped<FluentValidator>();
-        //services.AddScoped<FileImportService>();
-        //services.AddScoped<FileImporter>();
-
-        
 
     }
 }
